requirements.txt
google-api-python-client
pymongo
python-dotenv
textblob
pandas
matplotlib
seaborn